requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.0
gunicorn==20.1.0
djangorestframework==3.13.1
psycopg2==2.9.3
python-decouple==3.6
django-countries==7.3.2
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-jazzmin==2.5.0
dj-rest-auth==2.2.4
djangorestframework-simplejwt==5.1.0